requirements.txt
asgiref==3.10.0
click==8.3.0
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
h11==0.16.0
packaging==25.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
uvicorn==0.38.0
whitenoise==6.11.0
